requirements.txt
numpy
pandas
matplotlib
python-dotenv
mysql-connector-python
pymysql
scikit-learn
cryptography
seaborn
catboost
xgboost

# -e .